fixed deleted files after return

+++ routes/index.js
@@ -2,24 +2,27 @@ var express = require('express');
 var router = express.Router();
 
 var multer = require('multer');
-var upload = multer({ dest: 'uploads/' });
+var upload = multer({ dest: 'uploads/', limits:{ fileSize:100000 }, });
 var fs = require('fs-extra');
 
 /* GET home page. */
 router.get('/', function(req, res, next) {
-  res.render('index', { title: 'Express' });
+  /*fs.remove('./uploads/', err => {
+  });*/
+  return res.render('index');
+
 });
 
 router.post('/', upload.single('file'), function(req, res, next) {
-  console.log(req.file);
-  res.json({ size: req.file.size });
-
-  fs.remove('./uploads/', err => {
-  if (err) return console.error(err)
-
-  console.log('success!') // I just deleted my entire HOME directory.
-})
-
+  const { size } = req.file;
+  res.json({ size });
+  if(file){
+      fs.remove('./uploads/', err => {
+        if(err) { console.log("Errore"); }
+        else { console.log("succesfully deleted. "); }
+      });
+  }
+  return;
 });
 
 module.exports = router;
